Split index ts into two files
