+++ public/ar.html
@@ -13,6 +13,6 @@
     </head>
     <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
-        <script type="module" src="_dist_/index.js"></script>
+        <script type="module" src="_dist_/indexAR.js"></script>
     </body>
 </html>
