+++ public/index.html
@@ -121,6 +121,6 @@
                 </button>
             </div>
         </form>
-        <script type="module" src="_dist_/index.js"></script>
+        <script type="module" src="_dist_/indexNormal.js"></script>
     </body>
 </html>
